requirements.txt
argon2-cffi==20.1.0
attrs==19.3.0
bcrypt==3.1.7
blessed==1.17.9
boto3==1.14.43
botocore==1.17.43
cached-property==1.5.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
dj-database-url==0.5.0
Django==3.0.3
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-smartfields==1.1.1
django-storages==1.9.1
django-taggit==1.3.0
docker==4.3.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Faker==4.1.1
future==0.16.0
gunicorn==20.0.4
idna==2.7
importlib-metadata==1.7.0
jmespath==0.10.0
jsonschema==3.2.0
numpy==1.19.1
pandas==1.1.0
paramiko==2.7.1
pathspec==0.5.9
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.16.0
pytesseract==0.3.5
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
s3transfer==0.3.3
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.2.0
zipp==3.1.0
